Feature estates

+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/queries/GetPropertyByPriceRangeQuery.java
@@ -1,6 +1,4 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.queries;
 
-import java.math.BigDecimal;
-
-public record GetPropertyByPriceRangeQuery(BigDecimal minPrice, BigDecimal maxPrice) {
+public record GetPropertyByPriceRangeQuery(Long minPrice, Long maxPrice) {
 }
