+++ src/main/java/pe/edu/upc/micasita/Micasitaplataform/Property/domain/model/queries/GetPropertyBySizeRangeQuery.java
@@ -1,6 +1,5 @@
 package pe.edu.upc.micasita.Micasitaplataform.Property.domain.model.queries;
 
-import java.math.BigDecimal;
 
-public record GetPropertyBySizeRangeQuery(BigDecimal minSize, BigDecimal maxSize) {
+public record GetPropertyBySizeRangeQuery(Long minSize, Long maxSize) {
 }
